Add an environment concept
The environment concept allows passing of external data to pallet, and allows
project, service, or command line overrides of pallet algorithms and domain data
objects.

The :environment key is recognised by lift and converge, in the config.clj file
at the global and service scope. A project specific environment maybe be
specified in pallet.config/environment.