remove usage of deprecated method
The ComputeServiceUtils/getSupportedProviders class method is
deprecated and has been removed in jclouds/jclouds@dc4ca1efbd (which
is in the 1.6.0 branch).

This commit changes pallet.compute.jclouds/supported-providers to do
what ComputeServiceUtils/getSupportedProviders and
org.jclouds.rest.providers.Providers/getSupportedProvidersOfType used
to do.

Manual testing verified that for the current classpath, the old and
new implementations of the method both return the same set.